requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
cycler==0.11.0
DAWG-Python==0.7.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.4
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.12.2
docopt==0.6.2
docx==0.2.4
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.38.0
gunicorn==20.1.0
idna==3.4
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.22.4
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20220524
pickleshare==0.7.5
Pillow==9.3.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psutil==5.9.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymystem3==0.2.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2020.1
pywin32==304
pywinpty==2.0.9
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
scipy==1.9.3
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.3.0
stack-data==0.6.0
terminado==0.17.0
tinycss2==1.2.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
whitenoise==6.1.0
widgetsnbextension==3.6.1
